% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj_matrix2GREMLIN_format}
\alias{adj_matrix2GREMLIN_format}
\title{turn an adjacency matrix with a partition (functional groups) to a list of networks}
\usage{
adj_matrix2GREMLIN_format(adj_matrix, partition)
}
\description{
turn an adjacency matrix with a partition (functional groups) to a list of networks
}
