% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods.R
\name{coefficient_path}
\alias{coefficient_path}
\title{Extract the regularization path of a collection of fits}
\usage{
coefficient_path(x, type = c("partial_cor", "precision", "covariance"))
}
\arguments{
\item{x}{an object with class PLNnetworkfamily, i.e. an output from \code{\link{PLNnetwork}}}

\item{type}{a character, should we extract the path of covariance, precision or partial correlation coefficients ? Default is \code{partial_cor}.}
}
\value{
Send back a tibble/data.frame.
}
\description{
Extract the regularization path of a collection of fits
}
