% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_network.R
\name{estimate_network}
\alias{estimate_network}
\title{Estimation of the network structure (edges)}
\usage{
estimate_network(S, lambda, W = matrix(1, nrow(S), ncol(S)))
}
\arguments{
\item{S}{a d x d symmetric empirical covariance matrix}

\item{lambda}{postive real tuning the network sparsity}

\item{W}{a d x d matrix of weights}
}
\description{
The network is estimated with the adaptive Graphical-Lasso, where weigths are inversely
proportional to the probability of connection estimated by the SBM
}
