% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_precision.R
\name{estimate_precison}
\alias{estimate_precison}
\title{Adaptive graphical-Lasso}
\usage{
estimate_precison(S, Lambda, tau)
}
\arguments{
\item{S}{a d x d symmetric empirical covariance matrix}

\item{Lambda}{a K x K matrix of connectivity}

\item{tau}{a d x K of posterior probabilities of memberships}
}
\description{
Estimation of Stochastic Bloc Model under a Laplace emission law
}
\examples{
n <- 200; d <- 100; K <- 3
Lambda <- matrix(c(1, 2, 3, 2, 1, 4, 3, 4, 1), 3, 3)
alpha <- c(1/2, 1/4, 1/4)
K <- length(alpha)
sbm   <- rSBM_Laplace(d, Lambda, alpha)
cl <- igraph::V(sbm)$membership
Z <- matrix(0, d, K); Z[cbind(1:d, cl)] <- 1
X <- rmvnorm_SBM(n, sbm)
out <- estimate_precison(cov(X), 5 * n * Lambda, Z)
network <- out$wi; diag(network) <- 0
\dontrun{
image(Matrix(network[order(cl),order(cl)]))
}
}
