% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_structure.R
\name{estimate_structure}
\alias{estimate_structure}
\title{Estimation of the latent structure of the network}
\usage{
estimate_structure(adj_matrix, n_blocks,
  init_clustering = kmeans(adj_matrix, centers = n_blocks, nstart =
  10)$cl, threshold = 1e-05, max_iterates = 50)
}
\arguments{
\item{adj_matrix}{a symmetric weighted adjacency matrix}

\item{n_blocks}{integer, the number of blocks}

\item{init_clustering}{vector, an initial vector of memberships}

\item{threshold}{double, a convergence threshold}

\item{max_iterates}{integer,  the maximal number of itarates in the VEM algorithm}
}
\description{
The underlying network is assumed to be drawn from a Stochastic Bloc Model
with Laplace emission law for the edges. This function use a variational EM algorithm
to estimate such a model.
}
\examples{
Lambda <- matrix(c(1, 2, 3, 2, 1, 4, 3, 4, 1), 3, 3)
d <- 200
alpha <- c(1/2, 1/4, 1/4)
mySBM <- rSBM_Laplace(d, Lambda, alpha)
cl <- igraph::V(mySBM)$membership
adjMat <- mySBM \%>\% igraph::as_adj(attr = "weight")
vBlocks <- 1:5
out <- lapply(vBlocks, function(k) estimate_structure(adjMat, k))
}
