% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janine.R
\name{janine}
\alias{janine}
\title{Janine: Just Another Network Infernce mEthod}
\usage{
janine(
  data,
  penalty,
  n_blocks,
  epsilon = 0.001,
  max_iter = 20,
  penalize_diagonal = TRUE
)
}
\arguments{
\item{data}{a n x d matrix of multivariate Gaussian observation}

\item{penalty}{postive real tuning the network sparsity}

\item{n_blocks}{integer for the targetnumber of groups}

\item{epsilon}{convergence threshold, defaulkt to 1e-3}

\item{max_iter}{integer for the maximal number of iteration. Default to 20}
}
\description{
Iterate Adpative Graphical-Lasso with binary SBM estimation to recover the adaptive weights
}
