% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS-pkg.R
\docType{package}
\name{OLS-pkg}
\alias{OLS-pkg}
\alias{janine}
\alias{#}
\alias{pour}
\alias{associer}
\alias{au}
\alias{nom}
\alias{du}
\alias{package}
\alias{et}
\alias{avoir}
\alias{l'aide}
\alias{qui}
\alias{correspond}
\alias{à}
\alias{ça}
\alias{OLS-pkg-package}
\title{package janine}
\description{
Dans ce package, on trouve les fonctions trucs bidule et muche. Elles font ceci et cela. Les classes d'objets Robert et MItchoum font etc.
}
