% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods.R
\name{plot.janine_collection}
\alias{plot.janine_collection}
\title{Plot model selection criteria for a collection of fits}
\usage{
\method{plot}{janine_collection}(x, criteria = c("loglik", "BIC", "EBIC"), log.x = TRUE, ...)
}
\arguments{
\item{x}{an S3 object with class janine_collection}

\item{criteria}{vector of characters. The criteria to plot in c("loglik", "BIC", "EBIC").}

\item{log.x}{logical: should the x-axis be repsented in log-scale? Default is \code{TRUE}.}

\item{...}{use for S3 compatibility}
}
\description{
Plot model selection criteria for a collection of fits
}
