% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods.R
\name{plot.janine_fit}
\alias{plot.janine_fit}
\title{Extract and plot the network (partial correlation, support or inverse covariance) from a \code{janine_fit} object}
\usage{
\method{plot}{janine_fit}(
  x,
  type = c("partial_cor", "precision", "covariance", "support"),
  title = paste(match.arg(type), "matrix"),
  ...
)
}
\arguments{
\item{x}{an S3 object with class janine_fit}

\item{type}{character. Value of the weigth of the edges in the network, either "partial_cor" (partial correlation), "precision", "covariance" or "support" (binary). Default is \code{"partial_cor"}.}

\item{title}{an optional string for the title of the plot.}

\item{...}{Not used (S3 compatibility).}
}
\value{
Send back an invisible object (igraph or Matrix, depending on the output chosen) and optionaly displays a graph (via igraph or corrplot for large ones)
}
\description{
Extract and plot the network (partial correlation, support or inverse covariance) from a \code{janine_fit} object
}
