% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmvnorm_SBM}
\alias{rmvnorm_SBM}
\title{SBM-structured multivariate Gaussian observation}
\usage{
rmvnorm_SBM(n, SBMLaplace)
}
\arguments{
\item{n}{integer, the sample size}

\item{sbm}{an SBM, as outputed by the \code{\link{rSBMLaplace}} function}
}
\description{
Draw observations from a multivariate Gaussian distribution with SBM-structured precision matrix
}
\examples{
n <- 200; d <- 100; K <- 3
Lambda <- matrix(c(1, 2, 3, 2, 1, 4, 3, 4, 1), 3, 3)
alpha <- c(1/2, 1/4, 1/4)
K <- length(alpha)
sbm   <- rSBMLaplace(d, Lambda, alpha)
X <- rmvnorm_SBM(n, sbm)
}
