% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_laplace.R
\name{rSBMLaplace}
\alias{rSBMLaplace}
\title{rSBMLaplace}
\usage{
rSBMLaplace(p, Lambda, alpha = c(1))
}
\arguments{
\item{p}{number of nodes}

\item{Lambda}{a QxQ matrix of connectivity parameters}

\item{alpha}{a size-Q vector of mixture}
}
\value{
an igraph object: undirected weigthed graph with a attribute "membership" for the vertices
}
\description{
Generate a symmetric matrix under the stochastic block model
with a Laplace distribution as the emission law of the edges.
}
