% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods.R
\name{select_model}
\alias{select_model}
\title{Model selection from a collection of janine fits}
\usage{
select_model(x, crit = c("BIC", "EBIC", "loglik"))
}
\arguments{
\item{x}{an object with class janine_collection}

\item{crit}{a character for the criterion used to performed the selection. Either
"BIC", "EBIC" or "loglik". Default is \code{BIC}.}
}
\value{
Send back an object with class \code{janine_fit}
}
\description{
Model selection from a collection of janine fits
}
