% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEM_SBMLaplace.R
\name{VEM_SBM_laplace}
\alias{VEM_SBM_laplace}
\title{Estimation of a SBM Laplace}
\usage{
VEM_SBM_laplace(Omega, Q, cl.init = kmeans(Omega, centers = Q, nstart =
  10)$cl, eps = 1e-05, maxIter = 50)
}
\arguments{
\item{Omega}{a matrix: the data}

\item{Q}{an integer: the number of clusters}

\item{cl.init}{a initial clustering}

\item{eps}{double convergence threshold}

\item{maxIter}{integer the maximal number of itarates in the VEM algorithm}
}
\description{
This is the description of my function.
Estimation of Stochastic Bloc Model under a Laplace emission law
}
\examples{
Lambda <- matrix(c(1, 2, 3, 2, 1, 4, 3, 4, 1), 3, 3)
p <- 200
alpha <- c(1/2, 1/4, 1/4)
mySBM <- rSBMLaplace(p, Lambda, alpha)
cl <- igraph::V(mySBM)$membership
Omega <- mySBM \%>\% igraph::as_adj(attr = "weight")
vBlocks <- 1:5
out <- lapply(vBlocks, function(q) VEM_SBM_laplace(Omega, q))
}
