% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_block.R
\name{estimate_block}
\alias{estimate_block}
\title{Estimation of the latent block organisation of the network}
\usage{
estimate_block(adj_matrix, n_blocks, n_cores)
}
\arguments{
\item{adj_matrix}{a symmetric weighted adjacency matrix}

\item{n_blocks}{integer, the number of blocks}

\item{n_cores}{integer, the number of cores used for initializing the SBM exploration}
}
\description{
The underlying network is assumed to be drawn from a Stochastic Bloc Model.
This function uses a variational EM algorithm implemented in the package blockmodels to estimate such a model.
}
