% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janine.R
\name{janine}
\alias{janine}
\title{Janine: Just Another Network Infernce mEthod}
\usage{
janine(
  data,
  n_blocks,
  penalties = NULL,
  alpha = 0,
  control_optim = list(epsilon = 1e-04, max_iter = 20, trace = 1, n_cores = 4),
  control_penalties = list(min_ratio = 0.1, length = 20, diagonal = TRUE, weighted =
    TRUE)
)
}
\arguments{
\item{data}{a n x d matrix of multivariate Gaussian observation}

\item{n_blocks}{integer for the targetnumber of groups}

\item{penalties}{a vector of postive real number in decreasing order tuning the network sparsity. The default (NULL) generates
an hopefully appropriate collection of penalties.}

\item{alpha}{a positive scalar tuning the mixture between the weighted-sparse penlaty and the trace-Laplacian regularisation.}

\item{control_optim}{a list controling how the alternate optimization between adaptive graphical-Lasso and SBM is conducted
for each penalty level. Contains three entries: a convergence threshold \code{epsilon} (default to 1e-3),
a maximal number of iteration \code{max_iter} (default 20) and verbosity level \code{trace} (default 1).}

\item{control_penalties}{a list controling how \code{penalties} is generated, with three entries:
a double \code{min_ratio} (default 0.1), a integer \code{length} (default 25) and a logical \code{diagonal} (default FALSE)
indicating weither the diaognal should be penalized or not.}
}
\description{
Iterate Adaptive Graphical-Lasso with binary SBM estimation to recover the adaptive weights
}
\examples{
## Network settting
nNodes  <- 90
blockProp <- c(1/3, 1/3, 1/3)   # group proportions
nbBlock   <- length(blockProp) # number of blocks
connectParam <- diag(.4, nbBlock) + 0.01 # connectivity matrix: affiliation network
mySBM <- rggm::rSBM(nNodes, connectParam, blockProp)
Omega <- rggm::graph2prec(mySBM, cond_var = rep(1, nNodes), neg_prop = 0.5)
## Multivariate Gaussian Vector generation
n <- 500
X <- rggm::rmgaussian(n, means = rep(0, nNodes), solve(Omega))
## Network inference
out <- janine(X, n_blocks = 3, penalties = 0.1)
}
